Fix issue with unmanaged component registration after container restart. (Issue #889) (#902)

+++ trunk/workflow-manager/src/main/java/org/mitre/mpf/wfm/service/component/RemoveComponentServiceImpl.java
@@ -144,7 +144,9 @@ public class RemoveComponentServiceImpl implements RemoveComponentService {
         Path componentDir = getComponentTopLevelDir(registerModel.getJsonDescriptorPath());
 
         try {
-            MoreFiles.deleteRecursively(componentDir);
+            if (Files.exists(componentDir)) {
+                MoreFiles.deleteRecursively(componentDir);
+            }
             _componentStateService.removeComponent(registerModel.getComponentName());
         }
         catch (IOException ex) {
